Feat/#20: 통계 서비스,repository 작성
