+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/analytics/repository/GradeStatisticsRepository.java
@@ -1,0 +1,7 @@
+package com.example.cluvrapi.domain.analytics.repository;
+
+import com.example.cluvrapi.domain.analytics.entity.GradeStatistics;
+import com.example.cluvrapi.domain.common.repository.BaseRepository;
+
+public interface GradeStatisticsRepository extends BaseRepository<GradeStatistics, Long> {
+}
