+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/analytics/repository/PointStatisticsRepository.java
@@ -1,0 +1,7 @@
+package com.example.cluvrapi.domain.analytics.repository;
+
+import com.example.cluvrapi.domain.analytics.entity.PointStatistics;
+import com.example.cluvrapi.domain.common.repository.BaseRepository;
+
+public interface PointStatisticsRepository extends BaseRepository<PointStatistics, Long> {
+}
