+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/analytics/service/PointStatisticsService.java
@@ -1,0 +1,4 @@
+package com.example.cluvrapi.domain.analytics.service;
+
+public interface PointStatisticsService {
+}
