+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/analytics/service/PointStatisticsServiceImpl.java
@@ -1,0 +1,14 @@
+package com.example.cluvrapi.domain.analytics.service;
+
+import lombok.RequiredArgsConstructor;
+
+import org.springframework.stereotype.Service;
+
+import com.example.cluvrapi.domain.analytics.repository.PointStatisticsRepository;
+
+@Service
+@RequiredArgsConstructor
+public class PointStatisticsServiceImpl implements PointStatisticsService {
+	private final PointStatisticsRepository pointStatisticsRepository;
+
+}
